+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Calling `setText` on a BitmapText object will now recalculate its display origin values. Fix #3350 (thanks @migiyubi)
 * You can now pass an object to Loader.atlas, like you you can with images. Fix #3268 (thanks @TCatshoek)
 * The `onContextRestored` callback won't be defined any more unless the WebGL Renderer is in use in the following objects: BitmapMask, Static Tilemap, TileSprite and Text. This should allow those objects to now work in HEADLESS mode. Fix #3368 (thanks @16patsle)
+* The SetFrame method now has two optional arguments: `updateSize` and `updateOrigin` (both true by default) which will update the size and origin of the Game Object respectively. Fix #3339 (thanks @Jerenaux)
 
 ## Version 3.2.0 - Kaori - 5th March 2018